Conflicts:
	src/com/unus/smartrecorder/SRVoiceView.java

+++ src/com/unus/smartrecorder/SRDataSource.java
@@ -79,7 +79,7 @@ public class SRDataSource {
 	    return voices;
 	}
 	
-	public SRTagDb createTag(int voice_id, int type, String content, String tag_time){
+	public SRTagDb createTag(long voice_id, int type, String content, String tag_time){
     	ContentValues values = new ContentValues();
     	values.put(SRDbHelper.TAG_COLUMN_VOICE_ID, voice_id);
     	values.put(SRDbHelper.TAG_COLUMN_TYPE, type);
